First complete version, without testing, is now ready.
Working version of the code. Testing in progress.

Covering algorithm now only computes covering until 100% is reached for a genotype. This significantly increases performance. Also, the values of radius at which 100% is achieved is saved in file.

Output attributes corresponding to the 'Covering' analysis are now written to file in a pre-defined order.

Minor spelling correction and removal of dead code.

Code cleanup.

Further improvements towards bringing the code up to the PEP8 standard.

Implementing covering with comparisons within the GN.

First complete implementation of covering within GN. Requires testing.

Covering within GN seems to be working fine. The code is ready for thorough testing.

Bug fixes.

Output attributes corresponding to the 'Covering' analysis are now written to file in a pre-defined order.

Minor spelling correction and removal of dead code.

Covering within GN seems to be working fine. The code is ready for thorough testing.

Minor bug fixes.